+++ CONTRIBUTING.md
@@ -143,9 +143,9 @@ If you add a new feature, please also include appropriate tests to verify its in
 functionality. We try to keep our code coverage close to 100%.
 
 It is possible to limit the scope of testing to specific environments and files, for example, to
-only test transformations using the Python 3.7 environment use:
+only test event related functionality using the Python 3.7 environment use:
 ```bash
-tox -e py37 tests/transforms_test.py
+tox -e py37 tests/events
 ```
