+++ src/pymovements/events/detection/fill.py
@@ -26,7 +26,7 @@ import numpy as np
 
 from pymovements.events.events import EventDataFrame
 from pymovements.events.events import register_event_detection
-from pymovements.gaze.transforms import consecutive
+from pymovements.gaze.transforms_numpy import consecutive
 from pymovements.utils import checks
