+++ src/pymovements/events/detection/ivt.py
@@ -26,8 +26,8 @@ import numpy as np
 
 from pymovements.events.events import EventDataFrame
 from pymovements.events.events import register_event_detection
-from pymovements.gaze.transforms import consecutive
-from pymovements.gaze.transforms import norm
+from pymovements.gaze.transforms_numpy import consecutive
+from pymovements.gaze.transforms_numpy import norm
 from pymovements.utils import checks
 from pymovements.utils.filters import filter_candidates_remove_nans
