+++ src/pymovements/utils/filters.py
@@ -24,7 +24,7 @@ from __future__ import annotations
 
 import numpy as np
 
-from pymovements.gaze.transforms import consecutive
+from pymovements.gaze.transforms_numpy import consecutive
 
 
 def filter_candidates_remove_nans(
