+++ tests/events/detection/ivt_test.py
@@ -26,7 +26,7 @@ from polars.testing import assert_frame_equal
 
 from pymovements.events.detection.ivt import ivt
 from pymovements.events.events import EventDataFrame
-from pymovements.gaze.transforms import pos2vel
+from pymovements.gaze.transforms_numpy import pos2vel
 from pymovements.synthetic import step_function
