+++ tests/gaze/transforms/norm_test.py
@@ -59,6 +59,6 @@ import pymovements as pm
 )
 def test_norm_returns(columns, df, expected_series):
     result_df = df.select(
-        pm.gaze.transforms_pl.norm(columns=columns).alias('norm'),
+        pm.gaze.transforms.norm(columns=columns).alias('norm'),
     )
     assert_series_equal(result_df['norm'], expected_series, check_names=False)
