+++ tests/gaze/transforms/pos2acc_test.py
@@ -77,7 +77,7 @@ import pymovements as pm
 )
 def test_pos2acc_init_raises_error(kwargs, exception, msg_substrings):
     with pytest.raises(exception) as excinfo:
-        pm.gaze.transforms_pl.pos2acc(**kwargs)
+        pm.gaze.transforms.pos2acc(**kwargs)
 
     msg, = excinfo.value.args
     for msg_substring in msg_substrings:
@@ -101,7 +101,7 @@ def test_pos2acc_init_raises_error(kwargs, exception, msg_substrings):
 )
 def test_pos2acc_raises_error(kwargs, series, exception, msg_substrings):
     df = series.to_frame()
-    expression = pm.gaze.transforms_pl.pos2acc(**kwargs)
+    expression = pm.gaze.transforms.pos2acc(**kwargs)
 
     with pytest.raises(exception) as excinfo:
         df.select(expression)
@@ -213,6 +213,6 @@ def test_pos2acc_returns(kwargs, series, expected_df):
     df = series.to_frame()
 
     result_df = df.select(
-        pm.gaze.transforms_pl.pos2acc(**kwargs),
+        pm.gaze.transforms.pos2acc(**kwargs),
     )
     assert_frame_equal(result_df, expected_df.to_frame())
