+++ tests/gaze/transforms/pos2vel_test.py
@@ -120,7 +120,7 @@ import pymovements as pm
 )
 def test_pos2vel_init_raises_error(kwargs, exception, msg_substrings):
     with pytest.raises(exception) as excinfo:
-        pm.gaze.transforms_pl.pos2vel(**kwargs)
+        pm.gaze.transforms.pos2vel(**kwargs)
 
     msg, = excinfo.value.args
     for msg_substring in msg_substrings:
@@ -144,7 +144,7 @@ def test_pos2vel_init_raises_error(kwargs, exception, msg_substrings):
 )
 def test_pos2vel_raises_error(kwargs, series, exception, msg_substrings):
     df = series.to_frame()
-    expression = pm.gaze.transforms_pl.pos2vel(**kwargs)
+    expression = pm.gaze.transforms.pos2vel(**kwargs)
 
     with pytest.raises(exception) as excinfo:
         df.select(expression)
@@ -372,6 +372,6 @@ def test_pos2vel_returns(kwargs, series, expected_df):
     df = series.to_frame()
 
     result_df = df.select(
-        pm.gaze.transforms_pl.pos2vel(**kwargs),
+        pm.gaze.transforms.pos2vel(**kwargs),
     )
     assert_frame_equal(result_df, expected_df.to_frame())
