+++ tests/gaze/transforms/transforms_library_test.py
@@ -28,16 +28,16 @@ import pymovements as pm
 @pytest.mark.parametrize(
     ('transform_function', 'transform_name'),
     [
-        pytest.param(pm.gaze.transforms_pl.center_origin, 'center_origin', id='center_origin'),
-        pytest.param(pm.gaze.transforms_pl.downsample, 'downsample', id='downsample'),
-        pytest.param(pm.gaze.transforms_pl.norm, 'norm', id='norm'),
-        pytest.param(pm.gaze.transforms_pl.pix2deg, 'pix2deg', id='pix2deg'),
-        pytest.param(pm.gaze.transforms_pl.pos2acc, 'pos2acc', id='pos2acc'),
-        pytest.param(pm.gaze.transforms_pl.pos2vel, 'pos2vel', id='pos2vel'),
-        pytest.param(pm.gaze.transforms_pl.savitzky_golay, 'savitzky_golay', id='savitzky_golay'),
+        pytest.param(pm.gaze.transforms.center_origin, 'center_origin', id='center_origin'),
+        pytest.param(pm.gaze.transforms.downsample, 'downsample', id='downsample'),
+        pytest.param(pm.gaze.transforms.norm, 'norm', id='norm'),
+        pytest.param(pm.gaze.transforms.pix2deg, 'pix2deg', id='pix2deg'),
+        pytest.param(pm.gaze.transforms.pos2acc, 'pos2acc', id='pos2acc'),
+        pytest.param(pm.gaze.transforms.pos2vel, 'pos2vel', id='pos2vel'),
+        pytest.param(pm.gaze.transforms.savitzky_golay, 'savitzky_golay', id='savitzky_golay'),
     ],
 )
 def test_transform_registered(transform_function, transform_name):
-    assert transform_name in pm.gaze.transforms_pl.TransformLibrary.methods
-    assert pm.gaze.transforms_pl.TransformLibrary.get(transform_name) == transform_function
-    assert pm.gaze.transforms_pl.TransformLibrary.get(transform_name).__name__ == transform_name
+    assert transform_name in pm.gaze.transforms.TransformLibrary.methods
+    assert pm.gaze.transforms.TransformLibrary.get(transform_name) == transform_function
+    assert pm.gaze.transforms.TransformLibrary.get(transform_name).__name__ == transform_name
